refactor: move updatePeer to its own file
