+++ src/game-logic/reducers/updatePeer.test.js
@@ -1,0 +1,32 @@
+import { MAX_LATEST_PEER_MESSAGES } from '../../constants'
+
+import { updatePeer } from './updatePeer'
+
+describe('updatePeer', () => {
+  test('updates peer data', () => {
+    const { latestPeerMessages, peers } = updatePeer(
+      {
+        latestPeerMessages: [],
+        peers: { abc123: { foo: true } },
+      },
+      'abc123',
+      { foo: false }
+    )
+
+    expect(latestPeerMessages).toEqual([])
+    expect(peers).toEqual({ abc123: { foo: false } })
+  })
+
+  test('limits pendingPeerMessages', () => {
+    const { latestPeerMessages } = updatePeer(
+      {
+        latestPeerMessages: new Array(50).fill('message'),
+        peers: { abc123: { foo: true } },
+      },
+      'abc123',
+      { foo: false }
+    )
+
+    expect(latestPeerMessages).toHaveLength(MAX_LATEST_PEER_MESSAGES)
+  })
+})
